setup/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
coverage==7.6.0
cryptography==43.0.0
Django==5.0.7
django-allauth==0.63.6
django-widget-tweaks==1.5.0
gunicorn==22.0.0
idna==3.7
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
